testing live streams

+++ countingPolice.py
@@ -1231,24 +1231,41 @@ async def play_music(ctx,song):
     channel = song[2]
     runtime = song[3]
     author = song[4]
+    live = song[5]
     song = song[0]
+    if not live:
+        song_there = os.path.isfile("song.mp3")
 
-    song_there = os.path.isfile("song.mp3")
+        if song_there:
+            os.remove("song.mp3")
 
-    if song_there:
-        os.remove("song.mp3")
+        voice = ctx.guild.voice_client
 
-    voice = ctx.guild.voice_client
+        with youtube_dl.YoutubeDL(ydl_opts) as ydl:
+            ydl.download([song])
+        
+        
+        # await ctx.send(f"**Now Playing:** {title} - {channel} | {runtime}")
+        voice.play(discord.FFmpegPCMAudio(source="song.mp3"),after=lambda error: bot.loop.create_task(check_play_next(ctx)))
+        print("played audio...")
+        np_embed = discord.Embed(title="Now Playing", description=f"`{title}` requested by {author.mention}", value=f"Duration: {runtime}", color=bot_color)
+        np_embed.add_field(name=f"Duration: {runtime}", value=f"Channel: {channel}", inline=False)
+        await ctx.send(embed=np_embed)
+    else:
+        LIVE_YDL_OPTIONS = {'format': 'bestaudio/best', 'noplaylist':'True'}
+ 
+        FFMPEG_OPTIONS = {'before_options': '-reconnect 1 -reconnect_streamed 1 -reconnect_delay_max 5', 'options': '-vn'}
 
-    with youtube_dl.YoutubeDL(ydl_opts) as ydl:
-        ydl.download([song])
-    
-    
-    # await ctx.send(f"**Now Playing:** {title} - {channel} | {runtime}")
-    voice.play(discord.FFmpegPCMAudio(source="song.mp3"),after=lambda error: bot.loop.create_task(check_play_next(ctx)))
-    print("played audio...")
-    np_embed = discord.Embed(title="Now Playing", description=f"`{title}` requested by {author.mention}", value=f"Duration: {runtime}", color=bot_color)
-    await ctx.send(embed=np_embed)
+        voice = ctx.guild.voice_client
+        with youtube_dl.YoutubeDL(LIVE_YDL_OPTIONS) as ydl:
+            info = ydl.extract_info(song, download=False)
+            I_URL = info['formats'][0]['url']
+            source = await discord.FFmpegOpusAudio.from_probe(I_URL, **FFMPEG_OPTIONS)
+            voice.play(source, after = lambda error: bot.loop.create_task(check_play_next(ctx)))
+            print("played livestream...")
+            np_embed = discord.Embed(title="Now Playing", description=f"`{title}` request by {author.mention}", color = bot_color)
+            np_embed.add_field(name="Duration: :red_circle: LIVE", value=f"Channel: {channel}", inline=False)
+            await ctx.send(embed=np_embed)
 
 @bot.command(name="play", description="Plays a song in a voice channel", aliases=["p"])
 async def play(ctx, *args):
@@ -1291,11 +1308,18 @@ async def play(ctx, *args):
     if len(ytresults) == 0:
         await ctx.send("No results.")
         return
+    elif ytresults[0]['duration'] == 0:
+        song = "".join(("https://www.youtube.com", ytresults[0]['url_suffix']))
+        title = ytresults[0]['title']
+        channel = ytresults[0]["channel"]
+        runtime = 0
+        live = True
     else:
         song = "".join(("https://www.youtube.com", ytresults[0]["url_suffix"]))
         title = ytresults[0]["title"]
         channel = ytresults[0]["channel"]
         runtime = ytresults[0]["duration"]
+        live = False
 
     runtime_sec = col_to_sec(runtime)
 
@@ -1307,7 +1331,7 @@ async def play(ctx, *args):
 
     if voice:
         if voice.is_playing():
-            music_queue.append((song, title, channel, runtime, ctx.author))
+            music_queue.append((song, title, channel, runtime, ctx.author, live))
             total_runtime = 0
             for song in music_queue[1:]:
                 total_runtime += col_to_sec(song[3])
@@ -1318,10 +1342,10 @@ async def play(ctx, *args):
             await ctx.send(embed=queueadd_embed)
             return
         else:
-            await play_music(ctx, (song,title,channel, runtime, ctx.author))
+            await play_music(ctx, (song,title,channel, runtime, ctx.author, live))
     else:
         await uservoice.channel.connect()
-        await play_music(ctx,(song,title,channel, runtime, ctx.author))
+        await play_music(ctx,(song,title,channel, runtime, ctx.author, live))
 
 @bot.command(name="skip", description="Skips the currently playing song", aliases=["s"])
 async def skip(ctx):
